feat: type - base
You can now specify an external docker compose yaml file that will be
applied for the created instance.
This file will be included at 'up', 'down' and 'inspect' make targets.
Those file may be located anywhere on the file system and be very
specific to the project you're working on.